Get Real IP Address
If several IP Addresses are available, with the previous version it is not possible always to catch the real private IP address. For example on servers with running docker containers etc. Tested already on a couple of servers. The results are:

`ip addr | grep 'inet ' | tail -n1 | awk '{print $2}' | cut -f1  -d'/'`
```
172.17.0.1 # Docker container IP
```

While with:

`hostname -I | awk '{print $1}'`
```
172.16.100.200 # Real private IP
```