Update zero.sh
## Added

- Option to set Time Zone as variable /for more info see *Suggestion*/
- Option to set Phone Region as variable

## Removed

- Some trailing spaces

## Changed

- Get External IP Address
- Two `sed` statements that are setting the Time Zone in `/etc/php/8.0/cli/php.ini` and `/etc/php/8.0/fpm/php.ini` 
- Time Zone and Phone Region statements in `/var/www/nextcloud/config/config.php`

## Bugfix

- Uninstall Script: Escaping command substitution in Here Document creation of the "/home/$BENUTZERNAME/uninstall.sh" file which works now properly

## Suggestion

If you ask me, I would approach the Time Zone variable this way:
In the beginning of the script, it can run: `dpkg-reconfigure tzdata` interactively  to make sure proper Time Zone is set to the server. This one is a very frequent omission on many servers and they are left in some strange time zone. After that we can get the Time Zone from `/etc/timezone` file. If you wish I can do it. Just tell what is your preference.